Added support for templated assignees
Use case: Open a new issue if another issue is closed, under certain conditions.
Problem: We would want to retain the assignee from the original issue. Currently, this is not allowed.
Solution: This PR adds support for templated assignees. This way, `{{ mustache }}` case can be used and assignee can be made dynamic.